Complete flow refactor PR5 cleanup
