+++ docs/specs/roadmap.md
@@ -102,10 +102,10 @@
    - Phase 1 targets flow-viewer complexity.
    - Phase 2 targets application orchestration duplication.
    - Phase 3 targets domain conditional complexity.
-   - `flow-refactor` is the current concrete repository-native slice plan for
+   - `flow-refactor` completed the concrete repository-native slice plan for
      those phases, bundling diagnostics modularization, expanded-flow graph
-     separation, and `FlowContents.tsx` composition cleanup behind ordered
-     PR-sized slices.
+     separation, `FlowContents.tsx` composition cleanup, and final
+     naming/export cleanup behind ordered PR-sized slices.
    - Every slice must preserve desktop and web extension behavior.
 
 6. Make expanded flow layout deterministic.
