+++ src/ui-component/editor/ajsFlow/nodes/ConditionNode.tsx
@@ -22,7 +22,7 @@ import {
   handleKeyDownNavigateToTable,
 } from "./Utils";
 
-export type ConditionNode = Node<AjsNode, "condition">;
+type ConditionNode = Node<AjsNode, "condition">;
 type ConditionNodeProps = NodeProps<ConditionNode>;
 const ConditionNode: FC<ConditionNodeProps> = ({
   data,
