+++ src/ui-component/editor/ajsFlow/nodes/flowNodeDisplay.ts
@@ -1,0 +1,4 @@
+export const shouldRenderNodeComment = (
+  label?: string,
+  comment?: string,
+): boolean => !!comment && comment !== label;
